// src/common/enums/parcel-status.enum.ts
export enum ParcelStatus {
  PENDING_SUBMISSION = 'PENDING_SUBMISSION',
  SUBMITTED = 'SUBMITTED',
  SMS_SENT = 'SMS_SENT',
  SMS_FAILED = 'SMS_FAILED',
  PAYMENT_CONFIRMED = 'PAYMENT_CONFIRMED',
  RECEIPT_GENERATED = 'RECEIPT_GENERATED',
  ACTIVE = 'ACTIVE',
  IN_TRANSIT = 'IN_TRANSIT',
  ARRIVED = 'ARRIVED',
  DELIVERED = 'DELIVERED',
  FAILED_DELIVERY = 'FAILED_DELIVERY',
  RETURNED = 'RETURNED',
}